Move serializer KeyValuePair tests to dedicated file
