+++ src/libraries/System.Text.Json/tests/Serialization/CollectionTests/CollectionTests.Generic.Read.cs
@@ -963,167 +963,6 @@ namespace System.Text.Json.Serialization.Tests
         }
 
         [Fact]
-        public static void ReadSimpleKeyValuePairFail()
-        {
-            // Invalid form: no Value
-            Assert.Throws<JsonException>(() => JsonSerializer.Deserialize<KeyValuePair<string, int>>(@"{""Key"": 123}"));
-
-            // Invalid form: extra property
-            Assert.Throws<JsonException>(() => JsonSerializer.Deserialize<KeyValuePair<string, int>>(@"{""Key"": ""Key"", ""Value"": 123, ""Value2"": 456}"));
-
-            // Invalid form: does not contain both Key and Value properties
-            Assert.Throws<JsonException>(() => JsonSerializer.Deserialize<KeyValuePair<string, int>>(@"{""Key"": ""Key"", ""Val"": 123"));
-        }
-
-        [Fact]
-        public static void ReadListOfKeyValuePair()
-        {
-            List<KeyValuePair<string, int>> input = JsonSerializer.Deserialize<List<KeyValuePair<string, int>>>(@"[{""Key"": ""123"", ""Value"": 123},{""Key"": ""456"", ""Value"": 456}]");
-
-            Assert.Equal(2, input.Count);
-            Assert.Equal("123", input[0].Key);
-            Assert.Equal(123, input[0].Value);
-            Assert.Equal("456", input[1].Key);
-            Assert.Equal(456, input[1].Value);
-        }
-
-        [Fact]
-        public static void ReadKeyValuePairOfList()
-        {
-            KeyValuePair<string, List<int>> input = JsonSerializer.Deserialize<KeyValuePair<string, List<int>>>(@"{""Key"":""Key"", ""Value"":[1, 2, 3]}");
-
-            Assert.Equal("Key", input.Key);
-            Assert.Equal(3, input.Value.Count);
-            Assert.Equal(1, input.Value[0]);
-            Assert.Equal(2, input.Value[1]);
-            Assert.Equal(3, input.Value[2]);
-        }
-
-        [Theory]
-        [InlineData(@"{""Key"":""Key"", ""Value"":{""Key"":1, ""Value"":2}}")]
-        [InlineData(@"{""Key"":""Key"", ""Value"":{""Value"":2, ""Key"":1}}")]
-        [InlineData(@"{""Value"":{""Key"":1, ""Value"":2}, ""Key"":""Key""}")]
-        [InlineData(@"{""Value"":{""Value"":2, ""Key"":1}, ""Key"":""Key""}")]
-        public static void ReadKeyValuePairOfKeyValuePair(string json)
-        {
-            KeyValuePair<string, KeyValuePair<int, int>> input = JsonSerializer.Deserialize<KeyValuePair<string, KeyValuePair<int, int>>>(json);
-
-            Assert.Equal("Key", input.Key);
-            Assert.Equal(1, input.Value.Key);
-            Assert.Equal(2, input.Value.Value);
-        }
-
-        [Fact]
-        public static void ReadKeyValuePairWithNullValues()
-        {
-            {
-                KeyValuePair<string, string> kvp = JsonSerializer.Deserialize<KeyValuePair<string, string>>(@"{""Key"":""key"",""Value"":null}");
-                Assert.Equal("key", kvp.Key);
-                Assert.Null(kvp.Value);
-            }
-
-            {
-                KeyValuePair<string, object> kvp = JsonSerializer.Deserialize<KeyValuePair<string, object>>(@"{""Key"":""key"",""Value"":null}");
-                Assert.Equal("key", kvp.Key);
-                Assert.Null(kvp.Value);
-            }
-
-            {
-                KeyValuePair<string, SimpleClassWithKeyValuePairs> kvp = JsonSerializer.Deserialize<KeyValuePair<string, SimpleClassWithKeyValuePairs>>(@"{""Key"":""key"",""Value"":null}");
-                Assert.Equal("key", kvp.Key);
-                Assert.Null(kvp.Value);
-            }
-
-            {
-                KeyValuePair<string, KeyValuePair<string, string>> kvp = JsonSerializer.Deserialize<KeyValuePair<string, KeyValuePair<string, string>>>(@"{""Key"":""key"",""Value"":{""Key"":""key"",""Value"":null}}");
-                Assert.Equal("key", kvp.Key);
-                Assert.Equal("key", kvp.Value.Key);
-                Assert.Null(kvp.Value.Value);
-            }
-
-            {
-                KeyValuePair<string, KeyValuePair<string, object>> kvp = JsonSerializer.Deserialize<KeyValuePair<string, KeyValuePair<string, object>>>(@"{""Key"":""key"",""Value"":{""Key"":""key"",""Value"":null}}");
-                Assert.Equal("key", kvp.Key);
-                Assert.Equal("key", kvp.Value.Key);
-                Assert.Null(kvp.Value.Value);
-            }
-
-            {
-                KeyValuePair<string, KeyValuePair<string, SimpleClassWithKeyValuePairs>> kvp = JsonSerializer.Deserialize<KeyValuePair<string, KeyValuePair<string, SimpleClassWithKeyValuePairs>>>(@"{""Key"":""key"",""Value"":{""Key"":""key"",""Value"":null}}");
-                Assert.Equal("key", kvp.Key);
-                Assert.Equal("key", kvp.Value.Key);
-                Assert.Null(kvp.Value.Value);
-            }
-        }
-
-        [Fact]
-        public static void ReadClassWithNullKeyValuePairValues()
-        {
-            string json =
-                    @"{" +
-                        @"""KvpWStrVal"":{" +
-                            @"""Key"":""key""," +
-                            @"""Value"":null" +
-                        @"}," +
-                        @"""KvpWObjVal"":{" +
-                            @"""Key"":""key""," +
-                            @"""Value"":null" +
-                        @"}," +
-                        @"""KvpWClassVal"":{" +
-                            @"""Key"":""key""," +
-                            @"""Value"":null" +
-                        @"}," +
-                        @"""KvpWStrKvpVal"":{" +
-                            @"""Key"":""key""," +
-                            @"""Value"":{" +
-                                @"""Key"":""key""," +
-                                @"""Value"":null" +
-                            @"}" +
-                        @"}," +
-                        @"""KvpWObjKvpVal"":{" +
-                            @"""Key"":""key""," +
-                            @"""Value"":{" +
-                                @"""Key"":""key""," +
-                                @"""Value"":null" +
-                            @"}" +
-                        @"}," +
-                        @"""KvpWClassKvpVal"":{" +
-                            @"""Key"":""key""," +
-                            @"""Value"":{" +
-                                @"""Key"":""key""," +
-                                @"""Value"":null" +
-                            @"}" +
-                        @"}" +
-                    @"}";
-            SimpleClassWithKeyValuePairs obj = JsonSerializer.Deserialize<SimpleClassWithKeyValuePairs>(json);
-
-            Assert.Equal("key", obj.KvpWStrVal.Key);
-            Assert.Equal("key", obj.KvpWObjVal.Key);
-            Assert.Equal("key", obj.KvpWClassVal.Key);
-            Assert.Equal("key", obj.KvpWStrKvpVal.Key);
-            Assert.Equal("key", obj.KvpWObjKvpVal.Key);
-            Assert.Equal("key", obj.KvpWClassKvpVal.Key);
-            Assert.Equal("key", obj.KvpWStrKvpVal.Value.Key);
-            Assert.Equal("key", obj.KvpWObjKvpVal.Value.Key);
-            Assert.Equal("key", obj.KvpWClassKvpVal.Value.Key);
-
-            Assert.Null(obj.KvpWStrVal.Value);
-            Assert.Null(obj.KvpWObjVal.Value);
-            Assert.Null(obj.KvpWClassVal.Value);
-            Assert.Null(obj.KvpWStrKvpVal.Value.Value);
-            Assert.Null(obj.KvpWObjKvpVal.Value.Value);
-            Assert.Null(obj.KvpWClassKvpVal.Value.Value);
-        }
-
-        [Fact]
-        public static void Kvp_NullKeyIsFine()
-        {
-            KeyValuePair<string, string> kvp = JsonSerializer.Deserialize<KeyValuePair<string, string>>(@"{""Key"":null,""Value"":null}");
-            Assert.Null(kvp.Key);
-            Assert.Null(kvp.Value);
-        }
-
-        [Fact]
         public static void ReadSimpleTestClass_GenericCollectionWrappers()
         {
             SimpleTestClassWithGenericCollectionWrappers obj = JsonSerializer.Deserialize<SimpleTestClassWithGenericCollectionWrappers>(SimpleTestClassWithGenericCollectionWrappers.s_json);
