+++ src/libraries/System.Text.Json/tests/Serialization/CollectionTests/CollectionTests.Generic.Write.cs
@@ -751,117 +751,6 @@ namespace System.Text.Json.Serialization.Tests
         }
 
         [Fact]
-        public static void WritePrimitiveKeyValuePair()
-        {
-            KeyValuePair<string, int> input = new KeyValuePair<string, int>("Key", 123) ;
-
-            string json = JsonSerializer.Serialize(input);
-            Assert.Equal(@"{""Key"":""Key"",""Value"":123}", json);
-        }
-
-        [Fact]
-        public static void WriteListOfKeyValuePair()
-        {
-            List<KeyValuePair<string, int>> input = new List<KeyValuePair<string, int>>
-            {
-                new KeyValuePair<string, int>("123", 123),
-                new KeyValuePair<string, int>("456", 456)
-            };
-
-            string json = JsonSerializer.Serialize(input);
-            Assert.Equal(@"[{""Key"":""123"",""Value"":123},{""Key"":""456"",""Value"":456}]", json);
-        }
-
-        [Fact]
-        public static void WriteKeyValuePairOfList()
-        {
-            KeyValuePair<string, List<int>> input = new KeyValuePair<string, List<int>>("Key", new List<int> { 1, 2, 3 });
-
-            string json = JsonSerializer.Serialize(input);
-            Assert.Equal(@"{""Key"":""Key"",""Value"":[1,2,3]}", json);
-        }
-
-        [Fact]
-        public static void WriteKeyValuePairOfKeyValuePair()
-        {
-            KeyValuePair<string, KeyValuePair<string, int>> input = new KeyValuePair<string, KeyValuePair<string, int>>(
-                "Key", new KeyValuePair<string, int>("Key", 1));
-
-            string json = JsonSerializer.Serialize(input);
-            Assert.Equal(@"{""Key"":""Key"",""Value"":{""Key"":""Key"",""Value"":1}}", json);
-        }
-
-        [Fact]
-        public static void WriteKeyValuePairWithNullValues()
-        {
-            {
-                KeyValuePair<string, string> kvp = new KeyValuePair<string, string>("key", null);
-                Assert.Equal(@"{""Key"":""key"",""Value"":null}", JsonSerializer.Serialize(kvp));
-            }
-
-            {
-                KeyValuePair<string, object> kvp = new KeyValuePair<string, object>("key", null);
-                Assert.Equal(@"{""Key"":""key"",""Value"":null}", JsonSerializer.Serialize(kvp));
-            }
-
-            {
-                KeyValuePair<string, SimpleClassWithKeyValuePairs> kvp = new KeyValuePair<string, SimpleClassWithKeyValuePairs>("key", null);
-                Assert.Equal(@"{""Key"":""key"",""Value"":null}", JsonSerializer.Serialize(kvp));
-            }
-
-            {
-                KeyValuePair<string, KeyValuePair<string, string>> kvp = new KeyValuePair<string, KeyValuePair<string, string>>("key", new KeyValuePair<string, string>("key", null));
-                Assert.Equal(@"{""Key"":""key"",""Value"":{""Key"":""key"",""Value"":null}}", JsonSerializer.Serialize(kvp));
-            }
-
-            {
-                KeyValuePair<string, KeyValuePair<string, object>> kvp = new KeyValuePair<string, KeyValuePair<string, object>>("key", new KeyValuePair<string, object>("key", null));
-                Assert.Equal(@"{""Key"":""key"",""Value"":{""Key"":""key"",""Value"":null}}", JsonSerializer.Serialize(kvp));
-            }
-
-            {
-                KeyValuePair<string, KeyValuePair<string, SimpleClassWithKeyValuePairs>> kvp = new KeyValuePair<string, KeyValuePair<string, SimpleClassWithKeyValuePairs>>("key", new KeyValuePair<string, SimpleClassWithKeyValuePairs>("key", null));
-                Assert.Equal(@"{""Key"":""key"",""Value"":{""Key"":""key"",""Value"":null}}", JsonSerializer.Serialize(kvp));
-            }
-        }
-
-        // https://github.com/dotnet/runtime/issues/30388
-        [Fact]
-        public static void WriteClassWithNullKeyValuePairValues_NullWrittenAsEmptyObject()
-        {
-            var value = new SimpleClassWithKeyValuePairs()
-            {
-                KvpWStrVal = new KeyValuePair<string, string>("key", null),
-                KvpWObjVal = new KeyValuePair<string, object>("key", null),
-                KvpWClassVal = new KeyValuePair<string, SimpleClassWithKeyValuePairs>("key", null),
-                KvpWStrKvpVal = new KeyValuePair<string, KeyValuePair<string, string>>("key", new KeyValuePair<string, string>("key", null)),
-                KvpWObjKvpVal = new KeyValuePair<string, KeyValuePair<string, object>>("key", new KeyValuePair<string, object>("key", null)),
-                KvpWClassKvpVal = new KeyValuePair<string, KeyValuePair<string, SimpleClassWithKeyValuePairs>>("key", new KeyValuePair<string, SimpleClassWithKeyValuePairs>("key", null)),
-            };
-
-            string result = JsonSerializer.Serialize(value);
-
-            // Roundtrip to ensure serialize was correct.
-            value = JsonSerializer.Deserialize<SimpleClassWithKeyValuePairs>(result);
-            Assert.Equal("key", value.KvpWStrVal.Key);
-            Assert.Equal("key", value.KvpWObjVal.Key);
-            Assert.Equal("key", value.KvpWClassVal.Key);
-            Assert.Equal("key", value.KvpWStrKvpVal.Key);
-            Assert.Equal("key", value.KvpWObjKvpVal.Key);
-            Assert.Equal("key", value.KvpWClassKvpVal.Key);
-            Assert.Equal("key", value.KvpWStrKvpVal.Value.Key);
-            Assert.Equal("key", value.KvpWObjKvpVal.Value.Key);
-            Assert.Equal("key", value.KvpWClassKvpVal.Value.Key);
-
-            Assert.Null(value.KvpWStrVal.Value);
-            Assert.Null(value.KvpWObjVal.Value);
-            Assert.Null(value.KvpWClassVal.Value);
-            Assert.Null(value.KvpWStrKvpVal.Value.Value);
-            Assert.Null(value.KvpWObjKvpVal.Value.Value);
-            Assert.Null(value.KvpWClassKvpVal.Value.Value);
-        }
-
-        [Fact]
         public static void WriteGenericCollectionWrappers()
         {
             SimpleTestClassWithGenericCollectionWrappers obj1 = new SimpleTestClassWithGenericCollectionWrappers();
